yolo5/requirements.txt
pyTelegramBotAPI>=4.12.0
loguru>=0.7.0
requests>=2.31.0
flask>=2.3.2
matplotlib

pyyaml
loguru
boto3